refactor(site): move lifecycle settings under AI settings
