+++ site/src/pages/AISettingsPage/LifecyclePage/components/DurationField/DurationField.tsx
