+++ site/src/pages/AISettingsPage/LifecyclePage/components/LifecycleSettingLayout.tsx
@@ -2,8 +2,8 @@ import type { FC, FormEventHandler, ReactNode } from "react";
 import { Button } from "#/components/Button/Button";
 import { Spinner } from "#/components/Spinner/Spinner";
 import { Switch } from "#/components/Switch/Switch";
+import { TemporarySavedState } from "#/components/TemporarySavedState/TemporarySavedState";
 import { cn } from "#/utils/cn";
-import { TemporarySavedState } from "./TemporarySavedState";
 
 interface LifecycleSettingLayoutProps {
 	title: string;
