+++ site/src/pages/AISettingsPage/LifecyclePage/components/RetentionPeriodSettings.tsx
@@ -2,8 +2,8 @@ import { useFormik } from "formik";
 import type { FC } from "react";
 import * as Yup from "yup";
 import type * as TypesGen from "#/api/typesGenerated";
+import { useTemporarySavedState } from "#/components/TemporarySavedState/TemporarySavedState";
 import { DaysField, LifecycleSettingLayout } from "./LifecycleSettingLayout";
-import { useTemporarySavedState } from "./TemporarySavedState";
 
 interface MutationCallbacks {
 	onSuccess?: () => void;
