+++ site/src/pages/AISettingsPage/LifecyclePage/components/WorkspaceAutostopSettings.tsx
@@ -2,9 +2,9 @@ import { useFormik } from "formik";
 import type { FC } from "react";
 import * as Yup from "yup";
 import type * as TypesGen from "#/api/typesGenerated";
+import { useTemporarySavedState } from "#/components/TemporarySavedState/TemporarySavedState";
 import { DurationField } from "./DurationField/DurationField";
 import { LifecycleSettingLayout } from "./LifecycleSettingLayout";
-import { useTemporarySavedState } from "./TemporarySavedState";
 
 interface MutationCallbacks {
 	onSuccess?: () => void;
