+++ site/src/pages/AgentsPage/AgentsPageView.stories.tsx
@@ -34,7 +34,6 @@ import AgentSettingsCompactionPage from "./AgentSettingsCompactionPage";
 import AgentSettingsExperimentsPage from "./AgentSettingsExperimentsPage";
 import AgentSettingsGeneralPage from "./AgentSettingsGeneralPage";
 import AgentSettingsInstructionsPage from "./AgentSettingsInstructionsPage";
-import AgentSettingsLifecyclePage from "./AgentSettingsLifecyclePage";
 import AgentSettingsPage from "./AgentSettingsPage";
 import AgentSettingsSpendPage from "./AgentSettingsSpendPage";
 import { type AgentsOutletContext, AgentsPageView } from "./AgentsPageView";
@@ -211,7 +210,10 @@ const agentsRouting = {
 					element: <AgentSettingsInstructionsPage />,
 				},
 				{ path: "experiments", element: <AgentSettingsExperimentsPage /> },
-				{ path: "lifecycle", element: <AgentSettingsLifecyclePage /> },
+				{
+					path: "lifecycle",
+					element: <Navigate to="/ai/settings/lifecycle" replace />,
+				},
 				{ path: "admin", element: <AgentsRouteElement /> },
 				{ path: "agents", element: <AgentsRouteElement /> },
 				{ path: "spend", element: <AgentSettingsSpendPage now={fixedNow} /> },
